adjust run scripts

+++ run.sh
@@ -2,24 +2,28 @@
 
 trap "exit_func" INT
 
+if ! [ $(id -u) = 0 ]; then
+   echo "Root permissions required!"
+   exit 1
+fi
+
 exit_func() {
-    kill -9 $PID1
-    kill -9 $PID2
+    docker kill neo4j &> /dev/null
     exit
 }
 
+docker pull neo4j:3.5.11
+
 docker run --rm -t --publish=7687:7687 \
             --publish=7474:7474 --name neo4j \
-            --volume=$HOME/neo4j/data:/data \
-            --volume=$PWD/neo4j/plugins:/plugins \
-            --env NEO4J_dbms_connector_bolt_tls__level=DISABLED \
+            --volume=$HOME/neo4j_docker/data:/data \
+            --env NEO4J_dbms_connector_bolt_tls__level=OPTIONAL \
             --env NEO4J_AUTH=neo4j/neo3j \
-            --env 'NEO4JLABS_PLUGINS=["apoc"]' neo4j:3.5.11 &
+            --env 'NEO4JLABS_PLUGINS=["apoc"]' neo4j:3.5.11 > /dev/null &
 
-PID1=$!
+docker pull maxim615/coderadar
 
-./gradlew build
-docker build --tag coderadar:0.5 .
-docker run -it --rm -p 8080:8080 --name coderadar --network="bridge" coderadar:0.5
-PID2=$!
 
+sleep 20
+docker run -it --rm --name coderadar --network="host" maxim615/coderadar
+docker kill neo4j &> /dev/null
